# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Third Space Risk OS",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/venues/{venue_id}/incidents": {
      "get": {
        "summary": "List Incidents",
        "operationId": "list_incidents_api_venues__venue_id__incidents_get",
        "parameters": [
          {
            "name": "venue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Venue Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Incident"
                  },
                  "title": "Response List Incidents Api Venues  Venue Id  Incidents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Incident",
        "operationId": "create_incident_api_venues__venue_id__incidents_post",
        "parameters": [
          {
            "name": "venue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Venue Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentFlowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/incidents/{incident_id}/packets": {
      "get": {
        "summary": "List Incident Packets",
        "operationId": "list_incident_packets_api_incidents__incident_id__packets_get",
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Incident Packets Api Incidents  Incident Id  Packets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/packets/{packet_id}": {
      "get": {
        "summary": "Get Packet",
        "operationId": "get_packet_api_packets__packet_id__get",
        "parameters": [
          {
            "name": "packet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Packet Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Packet Api Packets  Packet Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/packets/{packet_id}/review-decisions": {
      "post": {
        "summary": "Create Review Decision",
        "operationId": "create_review_decision_api_packets__packet_id__review_decisions_post",
        "parameters": [
          {
            "name": "packet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Packet Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDecisionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Review Decision Api Packets  Packet Id  Review Decisions Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/packets/{packet_id}/audit-events": {
      "get": {
        "summary": "List Packet Audit Events",
        "operationId": "list_packet_audit_events_api_packets__packet_id__audit_events_get",
        "parameters": [
          {
            "name": "packet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Packet Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Packet Audit Events Api Packets  Packet Id  Audit Events Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/venues/{venue_id}/events/stream": {
      "post": {
        "summary": "Ingest Event Stream",
        "description": "High-volume ingestion endpoint. \nAccepts POS transactions, door scans, and camera metadata.",
        "operationId": "ingest_event_stream_api_venues__venue_id__events_stream_post",
        "parameters": [
          {
            "name": "venue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Venue Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/StreamEvent"
                },
                "title": "Events"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/venues/{venue_id}/risk-score": {
      "get": {
        "summary": "Get Venue Risk Score",
        "operationId": "get_venue_risk_score_api_venues__venue_id__risk_score_get",
        "parameters": [
          {
            "name": "venue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Venue Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Venue Risk Score Api Venues  Venue Id  Risk Score Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/venues/{venue_id}/quote": {
      "get": {
        "summary": "Get Venue Quote",
        "operationId": "get_venue_quote_api_venues__venue_id__quote_get",
        "parameters": [
          {
            "name": "venue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Venue Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Venue Quote Api Venues  Venue Id  Quote Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/venues/{venue_id}/compliance/{item_id}/upload": {
      "post": {
        "summary": "Upload Compliance Evidence",
        "operationId": "upload_compliance_evidence_api_venues__venue_id__compliance__item_id__upload_post",
        "parameters": [
          {
            "name": "venue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Venue Id"
            }
          },
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_compliance_evidence_api_venues__venue_id__compliance__item_id__upload_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Upload Compliance Evidence Api Venues  Venue Id  Compliance  Item Id  Upload Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ActionItem": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale"
          },
          "evidence_needed": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Needed"
          }
        },
        "type": "object",
        "required": [
          "title",
          "rationale",
          "evidence_needed"
        ],
        "title": "ActionItem"
      },
      "Body_upload_compliance_evidence_api_venues__venue_id__compliance__item_id__upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_compliance_evidence_api_venues__venue_id__compliance__item_id__upload_post"
      },
...
      "Citation": {
        "properties": {
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "excerpt": {
            "type": "string",
            "title": "Excerpt"
          }
        },
        "type": "object",
        "required": [
          "source_id",
          "source_type",
          "excerpt"
        ],
        "title": "Citation"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Incident": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "venue_id": {
            "type": "string",
            "title": "Venue Id"
          },
          "occurred_at": {
            "type": "string",
            "title": "Occurred At"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "reported_by": {
            "type": "string",
            "title": "Reported By"
          },
          "injury_observed": {
            "type": "boolean",
            "title": "Injury Observed"
          },
          "police_called": {
            "type": "boolean",
            "title": "Police Called"
          },
          "ems_called": {
            "type": "boolean",
            "title": "Ems Called"
          }
        },
        "type": "object",
        "required": [
          "id",
          "venue_id",
          "occurred_at",
          "location",
          "summary",
          "reported_by",
          "injury_observed",
          "police_called",
          "ems_called"
        ],
        "title": "Incident"
      },
      "IncidentCreate": {
        "properties": {
          "occurred_at": {
            "type": "string",
            "title": "Occurred At"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "reported_by": {
            "type": "string",
            "title": "Reported By"
          },
          "injury_observed": {
            "type": "boolean",
            "title": "Injury Observed"
          },
          "police_called": {
            "type": "boolean",
            "title": "Police Called"
          },
          "ems_called": {
            "type": "boolean",
            "title": "Ems Called"
          }
        },
        "type": "object",
        "required": [
          "occurred_at",
          "location",
          "summary",
          "reported_by",
          "injury_observed",
          "police_called",
          "ems_called"
        ],
        "title": "IncidentCreate"
      },
      "IncidentFlowResponse": {
        "properties": {
          "incident": {
            "$ref": "#/components/schemas/Incident"
          },
          "risk_signal": {
            "$ref": "#/components/schemas/RiskSignal"
          },
          "action_plan": {
            "items": {
              "$ref": "#/components/schemas/ActionItem"
            },
            "type": "array",
            "title": "Action Plan"
          },
          "claims_timeline": {
            "items": {
              "$ref": "#/components/schemas/TimelineEvent"
            },
            "type": "array",
            "title": "Claims Timeline"
          },
          "underwriting_memo": {
            "$ref": "#/components/schemas/UnderwritingMemo"
          }
        },
        "type": "object",
        "required": [
          "incident",
          "risk_signal",
          "action_plan",
          "claims_timeline",
          "underwriting_memo"
        ],
        "title": "IncidentFlowResponse"
      },
...
      "ReviewDecisionCreate": {
        "properties": {
          "reviewer_id": {
            "type": "string",
            "title": "Reviewer Id"
          },
          "decision": {
            "type": "string",
            "title": "Decision"
          },
          "override_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Override Reason"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "reviewer_id",
          "decision"
        ],
        "title": "ReviewDecisionCreate"
      },
      "RiskSignal": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          },
          "review_status": {
            "type": "string",
            "title": "Review Status"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "type": "array",
            "title": "Citations"
          }
        },
        "type": "object",
        "required": [
          "type",
          "severity",
          "confidence",
          "explanation",
          "review_status",
          "citations"
        ],
        "title": "RiskSignal"
      },
...
      "StreamEvent": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "event_type": {
            "type": "string",
            "title": "Event Type"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "event_type",
          "timestamp",
          "payload"
        ],
        "title": "StreamEvent"
      },
      "TimelineEvent": {
        "properties": {
          "at": {
            "type": "string",
            "title": "At"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "source": {
            "type": "string",
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "at",
          "label",
          "source"
        ],
        "title": "TimelineEvent"
      },
      "UnderwritingMemo": {
        "properties": {
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "open_questions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Open Questions"
          },
          "review_status": {
            "type": "string",
            "title": "Review Status"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "type": "array",
            "title": "Citations"
          }
        },
        "type": "object",
        "required": [
          "summary",
          "open_questions",
          "review_status",
          "citations"
        ],
        "title": "UnderwritingMemo"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}